Add a configurable prefix.
In an organization where lots of teams and apps
are using statsd, it is important to prefix
stats to keep them separate, but a pain to
repeat this prefix in every stat in your app.
This lets you define it (optionally) in one place.

The format I use in my work is "team.app.hostname."